# server/docs/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Access Control API",
		"description": "API REST para gerenciamento de usuários, salas e integrações IoT do sistema de controle de acesso.",
		"version": "1.0.0"
	},
	"components": {
		"securitySchemes": {
			"sessionCookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "auth_session",
				"description": "Sessão de usuário gerenciada pela camada Auth baseada em cookies."
			}
		},
		"schemas": {
			"Profile": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			}
		}
	},
	"paths": {
...
		"/rooms/": {
			"get": {
				"summary": "Listar salas e blocos",
				"tags": ["rooms"],
				"description": "Retorna as salas cadastradas, incluindo estado atual da porta e informações do bloco.",
				"security": [
					{
						"sessionCookie": []
					}
				],
				"responses": {
					"200": {
						"description": "Default Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"result": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"room": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"format": "uuid",
																"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
															},
															"name": {
																"type": "string"
															},
															"blockId": {
																"type": "string",
																"format": "uuid",
																"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
															},
															"isLocked": {
																"nullable": true,
																"type": "boolean"
															},
															"doorState": {
																"type": "string",
																"enum": ["OPEN", "CLOSED", "UNKNOWN"]
															},
															"lastStatusUpdateAt": {
																"nullable": true,
																"type": "string",
																"format": "date-time",
																"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
															},
															"createdAt": {
																"type": "string",
																"format": "date-time",
																"pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
															}
														},
														"required": [
															"id",
															"name",
															"blockId",
															"isLocked",
															"doorState",
															"lastStatusUpdateAt",
															"createdAt"
														],
														"additionalProperties": false
													},
													"block": {
														"type": "object",
														"properties": {
															"id": {
																"type": "string",
																"format": "uuid",
																"pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
															},
															"name": {
																"type": "string"
															}
														},
														"required": ["id", "name"],
														"additionalProperties": false
													}
												},
												"required": ["room", "block"],
												"additionalProperties": false
											}
										}
									},
									"required": ["result"],
									"additionalProperties": false,
									"example": {
										"result": [
											{
												"room": {
													"id": "cdea3efb-650d-4c8a-a9c0-8ee97d739f5c",
													"name": "Laboratório 101",
													"blockId": "1cf51c96-86a9-4fb3-b8e8-556a745d4423",
													"isLocked": true,
													"doorState": "CLOSED",
													"lastStatusUpdateAt": "2025-02-20T14:30:00.000Z",
													"createdAt": "2024-12-15T11:22:00.000Z"
												},
												"block": {
													"id": "1cf51c96-86a9-4fb3-b8e8-556a745d4423",
													"name": "Prédio Principal"
												}
											}
										]
									}
								}
							}
						}
					}
				}
			}
		},
...
		"/profiles/": {
			"get": {
				"summary": "Listar perfis",
				"tags": ["profiles"],
				"responses": {
					"200": {
						"description": "Lista de perfis",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"result": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/Profile"
											}
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"summary": "Criar perfil",
				"tags": ["profiles"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Perfil criado",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Profile"
								}
							}
						}
					}
				}
			}
		},
		"/users/{userId}/profiles": {
			"get": {
				"summary": "Listar perfis de um usuário",
				"tags": ["profiles"],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Perfis do usuário"
					}
				}
			},
			"post": {
				"summary": "Atribuir perfil ao usuário",
				"tags": ["profiles"],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"profileId": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"204": {
						"description": "Perfil atribuído"
					}
				}
			}
		},
		"/rooms/{roomId}/profiles": {
			"get": {
				"summary": "Listar perfis com acesso à sala",
				"tags": ["profiles"],
				"parameters": [
					{
						"in": "path",
						"name": "roomId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Perfis listados"
					}
				}
			},
			"post": {
				"summary": "Atribuir perfil à sala",
				"tags": ["profiles"],
				"parameters": [
					{
						"in": "path",
						"name": "roomId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"profileId": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"204": {
						"description": "Perfil atribuído à sala"
					}
				}
			}
		},
...
	},
...
}